chore: move package of common files
