+++ bm-common/src/main/java/org/benchmarker/bmagent/AgentStatus.java
@@ -1,4 +1,4 @@
-package org.benchmark.agent;
+package org.benchmarker.bmagent;
 
 public enum AgentStatus {
     READY, TESTING_START, TESTING, TESTING_FINISH, CANCELED, STOP_BY_ERROR
