+++ bm-common/src/main/java/org/benchmarker/bmcommon/BmCommonApplication.java
@@ -1,4 +1,4 @@
-package org.benchmark.bmcommon;
+package org.benchmarker.bmcommon;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
