+++ bm-common/src/main/java/org/benchmarker/bmcommon/dto/TemplateInfo.java
@@ -1,4 +1,4 @@
-package org.benchmark.dto;
+package org.benchmarker.bmcommon.dto;
 
 
 import java.util.Map;
