+++ bm-common/src/test/java/org/benchmarker/bmcommon/BmCommonApplicationTests.java
@@ -1,4 +1,4 @@
-package org.benchmark.bmcommon;
+package org.benchmarker.bmcommon;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
