+++ bm-controller/build.gradle
@@ -7,17 +7,18 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.4'
 }
 
-group = 'org.benchmark'
+group = 'org.benchmarker'
 version = rootProject.getVersion()
 
 // define exclude classes for Jacoco test coverage report
 def excludeJacocoTestCoverageReport = [
+        'org/benchmarker/bmcontroller/preftest/controller/PerftestController**',
         'org/benchmarker/bmcontroller/home/HomeController.class',
         'org/benchmarker/bmcontroller/user/controller/UserController.class',
-        'org/benchmarker/bmcontroller/BmControllerApplication.class',
+        'org/benchmarker/BmControllerApplication.class',
         'org/benchmarker/bmcontroller/login/controller/LoginController.class',
         'org/benchmarker/bmcontroller/common/util/NoOp.class',
-        'org/benchmarker/bmcontroller/admin/controller/AdminUserController.class',
+        'org/benchmarker/bmcontroller/controller/AdminUserController.class',
         'org/benchmarker/bmcontroller/user/controller/GroupController.class',
 ]
 
@@ -90,9 +91,9 @@ test {
     useJUnitPlatform()
 
     jacoco {
-        excludes = ["org/benchmarker/home/HomeController.class",
-                    "org/benchmarker/user/controller/UserController.class",
-                    "org/benchmarker/BmControllerApplication.class"]
+        excludes = ["org/benchmarker/bmcontroller/home/HomeController.class",
+                    "org/benchmarker/bmcontroller/user/controller/UserController.class",
+                    "org/benchmarker/bmcontroller/BmControllerApplication.class"]
     }
     finalizedBy jacocoTestReport
 }
